backend/requirements.txt
Flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
Werkzeug==2.3.7
openai==0.28.1 